+++ microapps/SceneViewCameraControllerApp/app/build.gradle.kts
@@ -1,6 +1,6 @@
 /*
  *
- *  Copyright 2023 Esri
+ *  Copyright 2024 Esri
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -28,11 +28,11 @@ secrets {
 }
 
 android {
-    namespace = "com.arcgismaps.toolkit.sceneviewtestapp"
+    namespace = "com.arcgismaps.toolkit.sceneviewcameracontrollerapp"
     compileSdk = libs.versions.compileSdk.get().toInt()
 
     defaultConfig {
-        applicationId ="com.arcgismaps.toolkit.sceneviewtestapp"
+        applicationId ="com.arcgismaps.toolkit.sceneviewcameracontrollerapp"
         minSdk = libs.versions.minSdk.get().toInt()
         targetSdk = libs.versions.compileSdk.get().toInt()
         versionCode = 1
