+++ microapps/SceneViewCameraControllerApp/app/src/main/java/com/arcgismaps/toolkit/sceneviewcameracontrollerapp/ui/theme/Theme.kt
@@ -1,6 +1,6 @@
 /*
  *
- *  Copyright 2023 Esri
+ *  Copyright 2024 Esri
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
